add task 2

+++ 0x00-python_variable_annotations/2-floor.py
@@ -7,15 +7,15 @@ Basic annotations - floor:
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     floor: function takes a float n as argument
     and returns the floor of the float.
 
-    Args: 
+    Args:
         n(float): n is the argument in float
 
-    Returns: 
+    Returns:
         Returns the floor of the float
     """
     return math.floor(n)
